wip implementing data file handling

+++ psycho/library/tools/size1.f
@@ -1,4 +1,4 @@
-inline Uint32 Tools::size(std::istream &stream)
+inline uint64_t Tools::size(std::istream &stream)
 {
     return stream.seekg(0, std::ios::end).tellg();
 }
