+++ psycho/library/tools/size2.f
@@ -1,4 +1,4 @@
-inline Uint32 Tools::size(std::ostream &stream)
+inline uint64_t Tools::size(std::ostream &stream)
 {
     return stream.seekp(0, std::ios::end).tellp();
 }
